# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/events/{id}:
    get:
      tags:
      - event-rest-controller
      summary: Get a single event by its id
      operationId: getEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventDTO"
    put:
      tags:
      - event-rest-controller
      summary: Update an event that already exists
      operationId: replaceEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Event"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventDTO"
    delete:
      tags:
      - event-rest-controller
      summary: Delete an event by its id
      operationId: deleteEvent
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/EventDTO"
  /api/v1/comments/{id}:
    get:
      tags:
      - comment-rest-controller
      summary: Get a comment by its id
      operationId: getCommentById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentDTO"
    put:
      tags:
      - comment-rest-controller
      summary: Update a comment by its id
      operationId: replaceComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
    delete:
      tags:
      - comment-rest-controller
      summary: Delete a comment by its id
      operationId: deleteComment
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
  /api/v1/bookings/{id}:
    get:
      tags:
      - bookings-rest-controller
      summary: Get a single booking by its id
      operationId: getBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingDTO"
    put:
      tags:
      - bookings-rest-controller
      summary: Update a booking that already exists
      operationId: updateBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingDTO"
    delete:
      tags:
      - bookings-rest-controller
      summary: Delete a booking by its id
      operationId: deleteBooking
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
...
  /api/v1/events/:
    get:
      tags:
      - event-rest-controller
      summary: Get all the events
      operationId: getAllEvents
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDTO"
    post:
      tags:
      - event-rest-controller
      summary: Create a new event
      operationId: createEvent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Event"
  /api/v1/comments/:
    get:
      tags:
      - comment-rest-controller
      summary: Get all comments
      operationId: getComments
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
    post:
      tags:
      - comment-rest-controller
      summary: Create a new comment
      operationId: createComment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CommentDTO"
    delete:
      tags:
      - comment-rest-controller
      summary: Delete all comments
      operationId: deleteAllComments
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
  /api/v1/bookings/:
    get:
      tags:
      - bookings-rest-controller
      summary: Get all the bookings done
      operationId: getAllBookings
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingDTO"
    post:
      tags:
      - bookings-rest-controller
      summary: Create a new booking
      operationId: createBooking
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingDTO"
        required: true
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BookingDTO"
...
  /api/v1/events/type/{type}:
    get:
      tags:
      - event-rest-controller
      summary: Get all the events of a certain type
      operationId: getAllEventsByType
      parameters:
      - name: type
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventDTO"
  /api/v1/comments/user/{userId}:
    get:
      tags:
      - comment-rest-controller
      summary: Get comments by user id
      operationId: getCommentByUserId
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
  /api/v1/comments/rate/{rate}:
    get:
      tags:
      - comment-rest-controller
      summary: Get comments by rate
      operationId: getCommentsByRate
      parameters:
      - name: rate
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommentDTO"
components:
  schemas:
...
    Event:
      type: object
      properties:
        id:
          type: integer
          format: int64
        description:
          type: string
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        title:
          type: string
        type:
          type: string
        category:
          type: string
    EventDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        type:
          type: string
        description:
          type: string
        category:
          type: string
    CommentDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        comentario:
          type: string
        eventId:
          type: integer
          format: int32
        rate:
          type: integer
          format: int32
    BookingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userName:
          type: string
        userEmail:
          type: string
        bussinesName:
          type: string
        capacity:
          type: integer
          format: int32
        eventDescript:
          type: string
        status:
          type: string
...